Python/app/requirements.txt
fastapi
uvicorn
cx_Oracle
python-dateutil
pydantic
